+++ src/converters/piral-blazor/src/interop.ts
@@ -1,6 +1,6 @@
 import { PiletMetadata } from 'piral-core';
 import { emitRenderEvent, emitNavigateEvent } from './events';
-import type { BlazorRootConfig } from './types';
+import type { BlazorRootConfig, WebAssemblyStartOptions } from './types';
 
 const wasmLib = 'Microsoft.AspNetCore.Components.WebAssembly';
 const coreLib = 'Piral.Blazor.Core';
@@ -55,7 +55,7 @@ function prepareForStartup() {
   }));
 }
 
-function createBlazorStarter(publicPath: string): () => Promise<BlazorRootConfig> {
+function createBlazorStarter(publicPath: string): (opts: WebAssemblyStartOptions) => Promise<BlazorRootConfig> {
   const root = document.body.appendChild(document.createElement('div'));
 
   root.style.display = 'none';
@@ -66,7 +66,7 @@ function createBlazorStarter(publicPath: string): () => Promise<BlazorRootConfig
     const originalBase = baseElement.href;
     baseElement.href = publicPath;
 
-    return () => {
+    return (opts) => {
       const navManager = window.Blazor._internal.navigationManager;
 
       //Overwrite to get NavigationManager in Blazor working, see https://github.com/smapiot/Piral.Blazor/issues/89
@@ -80,7 +80,7 @@ function createBlazorStarter(publicPath: string): () => Promise<BlazorRootConfig
 
       navManager.getBaseURI = () => originalBase;
 
-      return window.Blazor.start()
+      return window.Blazor.start(opts)
         .then(prepareForStartup)
         .then(({ capabilities, applyChanges }) => {
           baseElement.href = originalBase;
@@ -89,8 +89,8 @@ function createBlazorStarter(publicPath: string): () => Promise<BlazorRootConfig
     };
   }
 
-  return () =>
-    window.Blazor.start()
+  return (opts) =>
+    window.Blazor.start(opts)
       .then(prepareForStartup)
       .then(({ capabilities, applyChanges }) => [root, capabilities, applyChanges]);
 }
@@ -187,7 +187,11 @@ export function unloadBlazorPilet(id: string) {
   return window.DotNet.invokeMethodAsync(coreLib, 'UnloadPilet', id);
 }
 
-export function initialize(scriptUrl: string, publicPath: string) {
+export function initialize(scriptUrl: string, publicPath: string, opts: WebAssemblyStartOptions = {}) {
+  if (typeof opts.loadBootResource !== 'function') {
+    opts.loadBootResource = (type, name, url) => fetch(url, { method: 'GET', cache: 'no-cache' });
+  }
+
   return new Promise<BlazorRootConfig>((resolve, reject) => {
     const startBlazor = createBlazorStarter(publicPath);
     const script = document.createElement('script');
@@ -201,7 +205,7 @@ export function initialize(scriptUrl: string, publicPath: string) {
         emitNavigateEvent,
       });
 
-      startBlazor().then(resolve);
+      startBlazor(opts).then(resolve);
     };
 
     document.body.appendChild(script);
@@ -211,10 +215,10 @@ export function initialize(scriptUrl: string, publicPath: string) {
 export function createBootLoader(scriptUrl: string, satellites: Array<string>) {
   const publicPath = computePath();
 
-  return () => {
+  return (opts?: WebAssemblyStartOptions) => {
     if (typeof window.$blazorLoader === 'undefined') {
       // we load all satellite scripts before we initialize blazor
-      window.$blazorLoader = Promise.all(satellites.map(addScript)).then(() => initialize(scriptUrl, publicPath));
+      window.$blazorLoader = Promise.all(satellites.map(addScript)).then(() => initialize(scriptUrl, publicPath, opts));
     }
 
     return window.$blazorLoader;